refactor: moving `testhelpers` into `storage/internal`
